refactor(wework): extract cloud project context
